feat: scale to fit and center initially
The zoomable element should be always visible after initialization, hence
we have to scale either to the width or height, whatever results in the
smaller zoom-factor.
Also, the element should always be centered, if it doesn't have the same
dimensions as the container element. This includes the feature that it
always returns to the centered position upon zooming out.

To implement this, `getCurrentZoomCenter` was extracted, placed under
test and rewritten in Order to provide a zoom center that returns always
to the initial position/offset.