fix: don't treat doubletap as touchmove
Under android 7 with Google Chrome (but not FireFox o.O) an error
occurred that the taps from a zoom-out double tap also triggered an
touch-move event which caused the zoom-out animation from the doubleTap
to abort.

Hence a new field was introduced that is true if and only if we are in a
doubleTap and falsy otherwise.