fix: wait only 2 seconds to recenter loaded images
When using this library for cases in which the images are continually
replaced, then always resetting the zoom and offset is not desirable.

We cannot simply handle just the first load event, because cached images may
not cause any, or the library may be attached after the event fired.

Hence we choose to wait 2 seconds for any load event and afterwards only
adjust the initial offset silently.