feat: add options to allow dragging beyond the elements edge
As a side effect this allows to use the entire available space for
panning a zoom picture.

In this context, it also makes sense to be able to drag the element
even if unzoomed.

The zoomOutAnimation does not work for a zoomFaktor 1, because then
getCurrentZoomCenter cannot return a valid zoomCenter.