chore: switch to webpack
Switch to webpack has multiple strong advantages:

* splitting up the code into modules which are testable
* better development thanks to webpack's watch functionality
* more control over the files produced

This builds now another file with the `npm run build` command which
is not minified in case we want to use this library in another context
where we have another minifier.

This also extracts getCurrentZoomCenter and places it under test.

This finally adds a npm run watch command, making development much less
annoying.